FastAPISQLAlchamyStrawberryGraphQL/app/requirements.txt
fastapi==0.100.0
uvicorn==0.21.1
alembic==1.8.0
mysql-connector-python==8.0.29
asyncmy==0.2.8
sqlalchemy==1.4.42
jwt==1.3.1
bcrypt==3.1.7
cryptography==3.1.1
passlib==1.7.2
python-jose
python-multipart
async-exit-stack
async-generator
async-exit-stack
async-generator
uuid==1.30
strawberry-graphql==0.196.1
broadcaster==0.2.0
aiosqlite==0.19.0
websockets==11.0.3
aiosqlite==0.19.0
python-dotenv
httpx